Add further classes, modify domain model

+++ Software Construction/ucn.dmf83.sem1project.group4/src/ucn/dmf83/sem1project/group4/DomainLayer/Order.java
@@ -12,9 +12,9 @@ public class Order {
 	
 	private Map<Product,Integer> orderItems;
 	
-	private int orderPrice;
+	private double orderPrice;
 	private int discount;
-	private int finalprice;
+	private double finalprice;
 	
 	private boolean isPaid;
 }
